+++ fastcrypto/src/serde_helpers.rs
@@ -12,7 +12,7 @@ use serde_with::serde_as;
 use std::fmt;
 use std::fmt::{Debug, Display};
 
-use crate::error::FastCryptoError;
+use crate::error::{FastCryptoError, FastCryptoResult};
 use crate::{
     encoding::{Base64, Encoding},
     traits::{KeyPair, SigningKey, ToFromBytes, VerifyingKey},
@@ -277,7 +277,7 @@ impl Default for crate::bls12381::min_sig::BLS12381AggregateSignatureAsBytes {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::groups::bls12381::{G1Element, G1ElementAsBytes, G1_ELEMENT_BYTE_LENGTH};
+    use crate::groups::bls12381::{G1_ELEMENT_BYTE_LENGTH, G1Element, G1ElementAsBytes};
     use crate::groups::GroupElement;
     use schemars::schema_for;
 
@@ -328,4 +328,35 @@ mod tests {
             serde_json::to_string_pretty(&schema).unwrap()
         );
     }
+}
+
+/// Given a byte array of length `N * SIZE_IN_BYTES`, deserialize it into a vector of `N` elements
+/// of type `T`.
+pub fn deserialize_vector<const SIZE_IN_BYTES: usize, T: ToFromByteArray<SIZE_IN_BYTES>>(
+    bytes: &[u8],
+) -> FastCryptoResult<Vec<T>> {
+    if bytes.len() % SIZE_IN_BYTES != 0 {
+        return Err(FastCryptoError::InvalidInput);
+    }
+    bytes
+        .chunks(SIZE_IN_BYTES)
+        .map(|chunk| {
+            T::from_byte_array(
+                &chunk
+                    .try_into()
+                    .map_err(|_| FastCryptoError::InvalidInput)?,
+            )
+        })
+        .collect::<FastCryptoResult<Vec<T>>>()
+}
+
+/// Serialize a vector of elements of type T into a byte array by simply concatenating their binary
+/// representations.
+pub fn serialize_vector<const SIZE_IN_BYTES: usize, T: ToFromByteArray<SIZE_IN_BYTES>>(
+    elements: &[T],
+) -> Vec<u8> {
+    elements
+        .iter()
+        .flat_map(|e| e.to_byte_array().to_vec())
+        .collect()
 }
